+++ src/development/vortex/development/pipelines/__init__.py
@@ -1,5 +1,5 @@
-from vortex.development.core.pipelines.training_pipeline import TrainingPipeline
-from vortex.development.core.pipelines.validation_pipeline import PytorchValidationPipeline,IRValidationPipeline
-from vortex.development.core.pipelines.export_pipeline import GraphExportPipeline
-from vortex.development.core.pipelines.prediction_pipeline import PytorchPredictionPipeline, IRPredictionPipeline
-from vortex.development.core.pipelines.hypopt_pipeline import HypOptPipeline
+from .training_pipeline import TrainingPipeline
+from .validation_pipeline import PytorchValidationPipeline,IRValidationPipeline
+from .export_pipeline import GraphExportPipeline
+from .prediction_pipeline import PytorchPredictionPipeline, IRPredictionPipeline
+from .hypopt_pipeline import HypOptPipeline
