+++ src/development/vortex/development/pipelines/prediction_pipeline.py
@@ -11,12 +11,12 @@ import matplotlib.pyplot as plt
 import cv2
 import torch
 
-from vortex.development.core.factory import create_model, create_dataset, create_runtime_model
+from vortex.development.utils.factory import create_model, create_dataset, create_runtime_model
 from vortex.runtime import model_runtime_map
 from vortex.development.utils.visual import visualize_result
 from vortex.development.predictor import create_predictor, get_prediction_results
 from vortex.development.utils.common import check_and_create_output_dir
-from vortex.development.core.pipelines.base_pipeline import BasePipeline
+from vortex.development.pipelines.base_pipeline import BasePipeline
 
 __all__ = ['PytorchPredictionPipeline','IRPredictionPipeline']
